# ui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agent Control Server",
    "description": "Server component for Agent Control - policy-based control for AI agents.\n\n## Architecture\n\nThe system uses a hierarchical model:\n- **Agents**: AI systems that need control\n- **Policies**: Collections of control sets assigned to agents\n- **Control Sets**: Groups of related controls\n- **Controls**: Individual control configurations\n\n## Hierarchy\n\n```\nAgent → Policy → Control Set(s) → Control(s)\n```\n\n## Quick Start\n\n1. Register your agent with `/api/v1/agents/initAgent`\n2. Create controls with `/api/v1/controls` and configure them\n3. Create control sets and add controls to them\n4. Create a policy and add control sets to it\n5. Assign the policy to your agent\n6. Query agent's active controls with `/api/v1/agents/{agent_id}/controls`\n    ",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/agents/initAgent": {
      "post": {
        "tags": ["agents"],
        "summary": "Initialize or update an agent",
        "description": "Register a new agent or update an existing agent's tools and metadata.\n\nThis endpoint is idempotent:\n- If the agent name doesn't exist, creates a new agent\n- If the agent name exists with the same UUID, updates tool schemas\n- If the agent name exists with a different UUID, returns 409 Conflict\n\nTool versioning: When tool schemas change (arguments or output_schema),\na new version is created automatically.\n\nArgs:\n    request: Agent metadata and tool schemas\n    db: Database session (injected)\n\nReturns:\n    InitAgentResponse with created flag and active controls (if policy assigned)\n\nRaises:\n    HTTPException 409: Agent name exists with different UUID\n    HTTPException 500: Database error during creation/update",
        "operationId": "init_agent_api_v1_agents_initAgent_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/InitAgentRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Agent registration status with active controls",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InitAgentResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/agents/{agent_id}/controls": {
      "get": {
        "tags": ["agents"],
        "summary": "List agent's active controls",
        "description": "List all protection controls active for an agent.\n\nControls are inherited from all control sets in the agent's assigned policy.\nReturns an empty list if the agent has no policy.\n\nArgs:\n    agent_id: UUID of the agent\n    db: Database session (injected)\n\nReturns:\n    AgentControlsResponse with list of controls (empty if no policy)\n\nRaises:\n    HTTPException 404: Agent not found",
        "operationId": "list_agent_controls_api_v1_agents__agent_id__controls_get",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Agent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of controls from agent's policy",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentControlsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/controls": {
      "put": {
        "tags": ["controls"],
        "summary": "Create a new control",
        "description": "Create a new control with a unique name and empty data.\n\nControls define protection logic and can be added to control sets.\nUse the PUT /{control_id}/data endpoint to set control configuration.\n\nArgs:\n    request: Control creation request with unique name\n    db: Database session (injected)\n\nReturns:\n    CreateControlResponse with the new control's ID\n\nRaises:\n    HTTPException 409: Control with this name already exists\n    HTTPException 500: Database error during creation",
        "operationId": "create_control_api_v1_controls_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateControlRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Created control ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateControlResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Agent": {
        "properties": {
          "agent_id": {
            "type": "string",
            "format": "uuid",
            "title": "Agent Id",
            "description": "Unique identifier for the agent (UUID format)"
          },
          "agent_name": {
            "type": "string",
            "minLength": 1,
            "title": "Agent Name",
            "description": "Human-readable name for the agent"
          },
          "agent_description": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Agent Description",
            "description": "Optional description of the agent's purpose"
          },
          "agent_created_at": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Agent Created At",
            "description": "ISO 8601 timestamp when agent was created"
          },
          "agent_updated_at": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Agent Updated At",
            "description": "ISO 8601 timestamp when agent was last updated"
          },
          "agent_version": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Agent Version",
            "description": "Semantic version string (e.g. '1.0.0')"
          },
          "agent_metadata": {
            "anyOf": [
              { "additionalProperties": true, "type": "object" },
              { "type": "null" }
            ],
            "title": "Agent Metadata",
            "description": "Free-form metadata dictionary for custom properties"
          }
        },
        "type": "object",
        "required": ["agent_id", "agent_name"],
        "title": "Agent",
        "description": "Agent metadata for registration and tracking.\n\nAn agent represents an AI system that can be protected and monitored.\nEach agent has a unique ID and can have multiple tools registered with it.",
        "examples": [
          {
            "agent_description": "Handles customer inquiries and support tickets",
            "agent_id": "550e8400-e29b-41d4-a716-446655440000",
            "agent_metadata": {
              "environment": "production",
              "team": "support"
            },
            "agent_name": "customer-service-bot",
            "agent_version": "1.0.0"
          }
        ]
      },
      "AgentControlsResponse": {
        "properties": {
          "controls": {
            "items": { "$ref": "#/components/schemas/Control" },
            "type": "array",
            "title": "Controls",
            "description": "List of controls associated with the agent via its policy"
          }
        },
        "type": "object",
        "required": ["controls"],
        "title": "AgentControlsResponse"
      },
...
      "AgentTool": {
        "properties": {
          "tool_name": {
            "type": "string",
            "minLength": 1,
            "title": "Tool Name",
            "description": "Unique name for the tool"
          },
          "arguments": {
            "additionalProperties": true,
            "type": "object",
            "title": "Arguments",
            "description": "JSON schema describing tool input parameters"
          },
          "output_schema": {
            "additionalProperties": true,
            "type": "object",
            "title": "Output Schema",
            "description": "JSON schema describing tool output structure"
          }
        },
        "type": "object",
        "required": ["tool_name", "arguments", "output_schema"],
        "title": "AgentTool",
        "description": "Tool schema for agent capabilities.",
        "examples": [
          {
            "arguments": {
              "query": { "description": "Search query", "type": "string" }
            },
            "output_schema": {
              "results": { "items": { "type": "object" }, "type": "array" }
            },
            "tool_name": "search_knowledge_base"
          }
        ]
      },
...
      "Control": {
        "properties": {
          "id": { "type": "integer", "title": "Id" },
          "name": { "type": "string", "title": "Name" },
          "control": { "$ref": "#/components/schemas/ControlDefinition" }
        },
        "type": "object",
        "required": ["id", "name", "control"],
        "title": "Control",
        "description": "A control with identity and configuration.\n\nNote: Only fully-configured controls (with valid ControlDefinition)\nare returned from API endpoints. Unconfigured controls are filtered out."
      },
      "ControlAction": {
        "properties": {
          "decision": {
            "type": "string",
            "enum": ["allow", "deny", "warn", "log"],
            "title": "Decision",
            "description": "Action to take when control is triggered"
          }
        },
        "type": "object",
        "required": ["decision"],
        "title": "ControlAction",
        "description": "What to do when control matches."
      },
      "ControlDefinition": {
        "properties": {
          "description": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Description",
            "description": "Detailed description of the control"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "description": "Whether this control is active",
            "default": true
          },
          "local": {
            "type": "boolean",
            "title": "Local",
            "description": "If True, this control runs locally in the SDK. If False (default), it runs on the server.",
            "default": false
          },
          "applies_to": {
            "type": "string",
            "enum": ["llm_call", "tool_call"],
            "title": "Applies To",
            "description": "Which type of interaction this control applies to"
          },
          "check_stage": {
            "type": "string",
            "enum": ["pre", "post"],
            "title": "Check Stage",
            "description": "When to execute this control"
          },
          "selector": {
            "$ref": "#/components/schemas/ControlSelector",
            "description": "What data to select from the payload"
          },
          "evaluator": {
            "$ref": "#/components/schemas/EvaluatorConfig",
            "description": "How to evaluate the selected data"
          },
          "action": {
            "$ref": "#/components/schemas/ControlAction",
            "description": "What action to take when control matches"
          },
          "tags": {
            "items": { "type": "string" },
            "type": "array",
            "title": "Tags",
            "description": "Tags for categorization"
          }
        },
        "type": "object",
        "required": [
          "applies_to",
          "check_stage",
          "selector",
          "evaluator",
          "action"
        ],
        "title": "ControlDefinition",
        "description": "A control definition to evaluate agent interactions.\n\nThis model contains only the logic and configuration.\nIdentity fields (id, name) are managed by the database.",
        "examples": [
          {
            "action": { "decision": "deny" },
            "applies_to": "llm_call",
            "check_stage": "post",
            "description": "Block outputs containing US Social Security Numbers",
            "enabled": true,
            "evaluator": {
              "config": { "pattern": "\\b\\d{3}-\\d{2}-\\d{4}\\b" },
              "plugin": "regex"
            },
            "selector": { "path": "output" },
            "tags": ["pii", "compliance"]
          }
        ]
      },
...
      "ControlSelector": {
        "properties": {
          "path": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Path",
            "description": "Path to data using dot notation. Examples: 'input', 'output', 'arguments.query', 'context.user_id', 'tool_name', '*'",
            "default": "*"
          },
          "tool_names": {
            "anyOf": [
              { "items": { "type": "string" }, "type": "array" },
              { "type": "null" }
            ],
            "title": "Tool Names",
            "description": "Exact tool names this control applies to (ToolCall only)"
          },
          "tool_name_regex": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Tool Name Regex",
            "description": "RE2 pattern matched with search() against tool_name (ToolCall only)"
          }
        },
        "type": "object",
        "title": "ControlSelector",
        "description": "Selects data from payload and optionally scopes applicability by tool.\n\n- path: which slice of the payload to feed into the evaluator. Optional, defaults to \"*\"\n  meaning the entire payload object (ToolCall or LlmCall).\n- tool_names/tool_name_regex: optional applicability filters for ToolCall payloads.",
        "examples": [
          { "path": "output" },
          { "path": "arguments.query" },
          { "path": "context.user_id" },
          { "path": "input" },
          { "path": "*" },
          { "path": "arguments.dest", "tool_names": ["copy_file", "aws_cli"] },
          { "path": "output", "tool_name_regex": "^db_.*" }
        ]
      },
      "CreateControlRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Unique control name"
          }
        },
        "type": "object",
        "required": ["name"],
        "title": "CreateControlRequest"
      },
      "CreateControlResponse": {
        "properties": {
          "control_id": {
            "type": "integer",
            "title": "Control Id",
            "description": "Identifier of the created control"
          }
        },
        "type": "object",
        "required": ["control_id"],
        "title": "CreateControlResponse"
      },
...
      "EvaluatorConfig": {
        "properties": {
          "plugin": {
            "type": "string",
            "title": "Plugin",
            "description": "Plugin name or agent-scoped reference (agent:evaluator)",
            "examples": ["regex", "list", "my-agent:pii-detector"]
          },
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config",
            "description": "Plugin-specific configuration",
            "examples": [
              { "pattern": "\\d{3}-\\d{2}-\\d{4}" },
              { "logic": "any", "values": ["admin"] }
            ]
          }
        },
        "type": "object",
        "required": ["plugin", "config"],
        "title": "EvaluatorConfig",
        "description": "Evaluator configuration. See GET /plugins for available plugins and schemas.\n\nPlugin reference formats:\n- Built-in: \"regex\", \"list\"\n- Agent-scoped: \"my-agent:my-evaluator\" (validated in endpoint, not here)"
      },
...
      "EvaluatorSchema": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "description": "Unique evaluator name"
          },
          "config_schema": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config Schema",
            "description": "JSON Schema for evaluator config validation"
          },
          "description": {
            "anyOf": [
              { "type": "string", "maxLength": 1000 },
              { "type": "null" }
            ],
            "title": "Description",
            "description": "Optional description"
          }
        },
        "type": "object",
        "required": ["name"],
        "title": "EvaluatorSchema",
        "description": "Schema for a custom evaluator registered with an agent.\n\nCustom evaluators are PluginEvaluator classes deployed with the engine.\nThis schema is registered via initAgent for validation and UI purposes."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InitAgentRequest": {
        "properties": {
          "agent": {
            "$ref": "#/components/schemas/Agent",
            "description": "Agent metadata including ID, name, and version"
          },
          "tools": {
            "items": { "$ref": "#/components/schemas/AgentTool" },
            "type": "array",
            "title": "Tools",
            "description": "List of tools available to the agent"
          },
          "evaluators": {
            "items": { "$ref": "#/components/schemas/EvaluatorSchema" },
            "type": "array",
            "title": "Evaluators",
            "description": "Custom evaluator schemas for config validation"
          },
          "force_replace": {
            "type": "boolean",
            "title": "Force Replace",
            "description": "If true, replace corrupted agent data instead of failing. Use only when agent data is corrupted and cannot be parsed.",
            "default": false
          }
        },
        "type": "object",
        "required": ["agent"],
        "title": "InitAgentRequest",
        "description": "Request to initialize or update an agent registration.",
        "examples": [
          {
            "agent": {
              "agent_description": "Handles customer inquiries",
              "agent_id": "550e8400-e29b-41d4-a716-446655440000",
              "agent_name": "customer-service-bot",
              "agent_version": "1.0.0"
            },
            "evaluators": [
              {
                "config_schema": {
                  "properties": { "sensitivity": { "type": "string" } },
                  "type": "object"
                },
                "description": "Detects PII in text",
                "name": "pii-detector"
              }
            ],
            "tools": [
              {
                "arguments": { "query": { "type": "string" } },
                "output_schema": { "results": { "type": "array" } },
                "tool_name": "search_kb"
              }
            ]
          }
        ]
      },
      "InitAgentResponse": {
        "properties": {
          "created": {
            "type": "boolean",
            "title": "Created",
            "description": "True if agent was newly created, False if updated"
          },
          "controls": {
            "items": { "$ref": "#/components/schemas/Control" },
            "type": "array",
            "title": "Controls",
            "description": "Active protection controls for the agent (if policy assigned)"
          }
        },
        "type": "object",
        "required": ["created"],
        "title": "InitAgentResponse",
        "description": "Response from agent initialization."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    }
  }
}